made Node as outer class
