+++ src/datastructures/lrucache/LRUCache.java
@@ -3,21 +3,7 @@ package datastructures.lrucache;
 import java.util.HashMap;
 import java.util.Map;
 
-public class LRUCache {
-	class Node{
-		int key;
-		int value;
-		Node prev;
-		Node next;		
-		
-		public Node(int key, int value){
-			this.key = key;
-			this.value = value;
-			this.prev = null;
-			this.next = null;
-		}
-	}
-	
+public class LRUCache {	
 	int capcacity;
 	int count;
 	Node head;
