+++ src/datastructures/lrucache/Node.java
@@ -1,0 +1,15 @@
+package datastructures.lrucache;
+
+public class Node {
+	int key;
+	int value;
+	Node prev;
+	Node next;		
+	
+	public Node(int key, int value){
+		this.key = key;
+		this.value = value;
+		this.prev = null;
+		this.next = null;
+	}
+}
